Select NTLM9 password characters with div/mod, rather than more biased and/mod/shift.
This still leaves a small bias towards a part of the password set. There are 29 complete
repetitions of all possible passwords within the 64-bit index space, followed by a ~quarter
cycle which must be somehow biased as it can't cover everything.

This bias could be removed/reduced using larger indices, however doing so may be a poor
tradeoff. Assuming pessimistically that this is "wasted" would represent 0.92% of chain
entries being a waste. Because some work is repeated, these could lead to a 1.83%
performance loss at a chain length of 1,350,000. If avoiding the bias by using a larger
index, we would want to see no more than a 1.8% performance reduction to do so. Fast
and simple would be my choice in this case.

Fixes #2